Authorization added to all routes

+++ controllers/contactController.js
@@ -3,15 +3,15 @@ const Contact = require("../models/contactModel");
 
 //@desc Get all contacts
 //@route GET /api/contacts
-//@access public
+//@access private
 const getContacts = asyncHandler(async (req, res) => {
-    const contacts = await Contact.find();
+    const contacts = await Contact.find({ user_id: req.user.id });
     res.status(200).json(contacts);
 });
 
 //@desc Get contact
 //@route GET /api/contacts
-//@access public
+//@access private
 const getContact = asyncHandler(async(req, res) => {
     const contact = await Contact.findById(req.params.id);
     if(!contact) {
@@ -23,7 +23,7 @@ const getContact = asyncHandler(async(req, res) => {
 
 //@desc Create new contact
 //@route Post /api/contacts
-//@access public
+//@access private
 const createContact = asyncHandler(async(req, res) => {
     console.log("BODY ---->", req.body);
     const { name, email, phone } = req.body;
@@ -35,18 +35,24 @@ const createContact = asyncHandler(async(req, res) => {
         name,
         email,
         phone,
+        user_id: req.user.id,
     });
     res.status(201).json(contact);
 });
 
 //@desc Update a contact
 //@route PUT /api/contacts/:id
-//@access public
+//@access private
 const updateContact = asyncHandler(async(req, res) => {
     const contact = await Contact.findById(req.params.id);
     if(!contact) {
         res.status(404);
         throw new Error("Contact not found!!");
+    }
+
+    if(contact.user_id.toString() !== req.user.id) {
+        res.status(403);
+        throw new Error("Not authorized to update this contact!!");
     }
 
     const updatedContact = await Contact.findByIdAndUpdate(
@@ -59,12 +65,16 @@ const updateContact = asyncHandler(async(req, res) => {
 
 //@desc Delete a contacts
 //@route DELETE /api/contacts/:id
-//@access public
+//@access private
 const deleteContact = asyncHandler(async(req, res) => {
     const contact = await Contact.findById(req.params.id);
     if(!contact) {
         res.status(404);
         throw new Error("Contact not found!!");
+    }
+    if(contact.user_id.toString() !== req.user.id) {
+        res.status(403);
+        throw new Error("Not authorized to delete this contact!!");
     }
     await Contact.findByIdAndDelete(req.params.id);
     res.status(201).json({ message: `Delete contact for ${req.params.id}`});
